+++ packages/3d-globe-map/data/world.js
@@ -1,0 +1,2 @@
+export * from "canvas-globe/data/world";
+export { default } from "canvas-globe/data/world";