+++ packages/canvas-globe-web-component/index.js
@@ -1,0 +1,1 @@
+export { GeoGlobeElement, defineGeoGlobe } from "canvas-globe/element";